prettify, add instructions for bootstrap, add preamble feature
The preamble feature will ask the user if the commands
which will be executed are

The preamble feature will present the user with an optional note,
and the commands which will be executed if the user types "yes." If
the user types "no," the script will exit, otherwise it will keep
asking the user to type either `yes` or `no`. This feature can be
skipped with the `--skip-preamble` feature.

I prettified `README.md`'s formatting and added a note about using
the bootstrap.

I had to add `--skip-preamble` to the `.travis.yml` for automation
purposes (silent install).