[Docs] Use relative links to documents in the README
This patch replaces two absolute URLs in the README with relative
links instead.  For example, the link to the "Contributing" document
now simply links to `./CONTRIBUTING.md`.  The benefit to this is that
when people view the README on GitHub the links will point to the
version of those documents using the same branch as the README.  See
the link below for more information on why this is a good practice.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: https://help.github.com/articles/relative-links-in-readmes/