NEW contextsprite.py: redoing Walkabout, +more!
Add `contextsprite.py`, it is going to be a replacement
for the current Walkabout to more closely align with our
short term goals for the engine.

Rename `constants.Direction.cardinal()` to `cardinals()`.

Change `README.md` demo gif and fix GitHub license link.

This commit will not pass tests on Travis-CI. I have made
this commit so that my work may be reviewed. Some of it has
copied portions from `Walkabout`. I have not even ran this.
This branch is far from done or functional.

See Also:

  * http://hypatia.software/2016/01/10/clear-hypatia-engine-short-term-vision/
  * http://hypatia.software/2015/12/05/hypatia-engine-short-term-vision/