Redo resources, update tests, update actor
Lots of things have been changed to allow for
dictionaries instead of a `Resource` object.

The resource module was redone to primarily
be a function that returns a dictionary.

`@requires_pygame_direction` modified to actually
work as a decorator! Change respective references
to this decorator.

Add `from_files_dict` boilerplate to `ContextSprite`.